+++ providers/instance.rb
@@ -232,11 +232,11 @@ def load_service_definitions_and_keys(service_definitions)
             group group
             mode 00600
             sensitive true
-            notifies :run, 'bash[make_keystore_from_new_file]', :delayed
+            notifies :run, "bash[Make Keystore For #{host}]", :delayed
             not_if { ::File.exist?("#{ssl_dir}/#{host}.#{name}.crt") }
           end
 
-          bash 'make_keystore_from_new_file' do
+          bash "Make Keystore For #{host}" do
             cwd ssl_dir
             code <<-EOH
             keytool -import -noprompt -trustcacerts -alias #{host} -file #{ssl_dir}/#{host}.#{name}.crt -keystore truststore -srcstorepass #{keystore_password} -deststorepass #{keystore_password}